% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_season_col.R
\name{append_season_col}
\alias{append_season_col}
\title{Partitions dataframe into different Seasons based on a transition out of playoffs 
df -> df}
\usage{
append_season_col(df)
}
\description{
Partitions dataframe into different Seasons based on a transition out of playoffs 
df -> df
}

