% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrNum.R
\name{corrNum}
\alias{corrNum}
\title{Cor numeric}
\usage{
corrNum(season)
}
\arguments{
\item{season}{a year of interest}
}
\value{
correlation plot of the numeric variables of a given data set
}
\description{
Takes all the numeric columns in a dataset and returns a correlation plot
}
\examples{
corrNum()
}
