% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bball_data.R
\name{exp_pts}
\alias{exp_pts}
\title{Calculates total Expeted Points Scored of 2 Arbitrary Teams}
\usage{
exp_pts(teamA, teamB)
}
\description{
Calculates total Expeted Points Scored of 2 Arbitrary Teams
}

