% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bball_data.R
\name{get_full_rpm}
\alias{get_full_rpm}
\title{Fetch full Real Plus-Minus for all players (taken from espn)
Note. This is a snapshot as of the state of world when the function was called.}
\usage{
get_full_rpm()
}
\description{
Fetch full Real Plus-Minus for all players (taken from espn)
Note. This is a snapshot as of the state of world when the function was called.
}

