% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_bball_data.R
\name{get_srs}
\alias{get_srs}
\title{Retrieves current snapshot of team SRS data (taken from bball-ref)
team simple rating system that takes into account point differential & strength of schedule}
\usage{
get_srs()
}
\description{
Retrieves current snapshot of team SRS data (taken from bball-ref)
team simple rating system that takes into account point differential & strength of schedule
}

