% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_regular_season.R
\name{is_regular_season}
\alias{is_regular_season}
\title{Whether the nba game date falls within regular season. 
Games outside of this fall within the playoffs. Note there are lockout seasons not currently handled}
\usage{
is_regular_season(gm_date)
}
\description{
Whether the nba game date falls within regular season. 
Games outside of this fall within the playoffs. Note there are lockout seasons not currently handled
}

