% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lebronPPG.R
\name{lebronPPG}
\alias{lebronPPG}
\title{Points Per Game}
\usage{
lebronPPG(season)
}
\arguments{
\item{season}{a year of interest.}
}
\value{
points per game (taking the total points (PTS) scored in a season divided by the total games (G) played in that season)
}
\description{
This function will return the points per game for player "LeBron James" for any of the seasons that he has played (2004 - 2017).
}
\examples{
lebronPPG()
}
