% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba}
\alias{nba}
\title{Dataset that contains aggregate statistics of 67 NBA seasons}
\format{
A dataframe with 53 variables over 24,691 rows
}
\source{
https://www.kaggle.com/drgilermo/nba-players-stats/data?select=Seasons_Stats.csv
}
\usage{
nba
}
\description{
Contains 53 different player statistics (e.g. PPG) from NBA seasons
}
\examples{
data(nba)
}
\keyword{datasets}
