% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldestPlayer.R
\name{oldestPlayer}
\alias{oldestPlayer}
\title{Oldest player}
\usage{
oldestPlayer(season)
}
\arguments{
\item{season}{a year of interest}
}
\value{
oldest player for a given year
}
\description{
Returns the oldest player for any given season. If there are multiple older players, then the first alphabetically listed player will be returned.
}
\examples{
oldestPlayer()
}
